refactor(D-168 A4): extract chat_stream_engine from routes.py + late-binding lockstep
- chat_stream_engine.py: _stream_chat_langgraph + _run_chat_langgraph (verbatim)
  + active_background_tasks module state; imports flow one-way from siblings
- routes.py re-imports the engines; handlers call them unchanged
- manifest: API_SOURCE_FILES += chat_stream_engine.py
- late-binding lockstep (patch the module where the caller lives):
  * test_orchestrator_chat_stategraph: WS _persist patches -> chat_stream_engine;
    FakeGraph implements astream (ISS-STREAM-002 contract) + STATE_DIR -> tmp_path
    => the 3 historically-deselected stategraph tests are green again
  * ci.yml: deselect list SHRANK by 3 (D-105 rule: lists shrink only)
  * test_outbox_relay_control_plane + test_orchestrator_admin_tool_security:
    get_settings patches -> identity_access (guard moved in A2)

routes.py: 3,312 -> 1,561 lines total (-53%). Fix: chat_stream_engine imports
_safe_conversation_id (was a routes-global resolution).
