+++ tests/microservices/test_orchestrator_chat_stategraph.py
@@ -6,7 +6,7 @@ import pytest
 from fastapi.testclient import TestClient
 
 from microservices.orchestrator_service.main import app
-from microservices.orchestrator_service.src.api import routes
+from microservices.orchestrator_service.src.api import chat_stream_engine, routes
 
 
 class _FakeTimelineEvent:
@@ -56,8 +56,10 @@ class _FakeLangGraphService:
         return _FakeRunData()
 
 
-def test_chat_http_messages_uses_stategraph(monkeypatch) -> None:
+def test_chat_http_messages_uses_stategraph(monkeypatch, tmp_path) -> None:
     """يتأكد أن POST /api/chat/messages يعيد استجابة موحدة من مسار StateGraph."""
+
+    monkeypatch.setenv("STATE_DIR", str(tmp_path))  # telemetry evidence → tmp
 
     class FakeGraph:
         async def astream_events(self, *args, **kwargs):
@@ -66,6 +68,10 @@ def test_chat_http_messages_uses_stategraph(monkeypatch) -> None:
                 "name": "LangGraph",
                 "data": {"output": {"final_response": "stategraph-response"}},
             }
+
+        async def astream(self, *args, **kwargs):
+            # ISS-STREAM-002: المسار الحي يستهلك astream(stream_mode=["custom","updates"])
+            yield ("updates", {"synthesizer": {"final_response": "stategraph-response"}})
 
     def _fake_create_unified_graph():
         return FakeGraph()
@@ -102,8 +108,10 @@ import jwt
 from microservices.orchestrator_service.src.core.config import get_settings
 
 
-def test_chat_ws_customer_uses_stategraph(monkeypatch) -> None:
+def test_chat_ws_customer_uses_stategraph(monkeypatch, tmp_path) -> None:
     """يتأكد أن WS العميل يمر عبر نفس مسار StateGraph ويرجع route_id الصحيح."""
+
+    monkeypatch.setenv("STATE_DIR", str(tmp_path))  # telemetry evidence → tmp
 
     class FakeGraph:
         async def astream_events(self, *args, **kwargs):
@@ -112,6 +120,10 @@ def test_chat_ws_customer_uses_stategraph(monkeypatch) -> None:
                 "name": "LangGraph",
                 "data": {"output": {"final_response": "Fake Graph WS Result"}},
             }
+
+        async def astream(self, *args, **kwargs):
+            # ISS-STREAM-002: المحرك يستهلك astream(stream_mode=["custom","updates"])
+            yield ("updates", {"synthesizer": {"final_response": "Fake Graph WS Result"}})
 
     def _fake_create_unified_graph():
         return FakeGraph()
@@ -129,7 +141,9 @@ def test_chat_ws_customer_uses_stategraph(monkeypatch) -> None:
     async def fake_persist_assistant_message(**kwargs):
         pass
 
-    monkeypatch.setattr(routes, "_persist_assistant_message", fake_persist_assistant_message)
+    monkeypatch.setattr(
+        chat_stream_engine, "_persist_assistant_message", fake_persist_assistant_message
+    )
 
     token = jwt.encode({"sub": "1", "user_id": 1}, get_settings().SECRET_KEY, algorithm="HS256")
     with TestClient(app).websocket_connect(f"/api/chat/ws?token={token}") as ws:
@@ -149,8 +163,10 @@ def test_chat_ws_customer_uses_stategraph(monkeypatch) -> None:
                 break
 
 
-def test_chat_ws_admin_uses_stategraph(monkeypatch) -> None:
+def test_chat_ws_admin_uses_stategraph(monkeypatch, tmp_path) -> None:
     """يتأكد أن WS الإداري يستخدم StateGraph ويرجع route_id الإداري."""
+
+    monkeypatch.setenv("STATE_DIR", str(tmp_path))  # telemetry evidence → tmp
 
     class FakeGraph:
         async def astream_events(self, *args, **kwargs):
@@ -159,6 +175,10 @@ def test_chat_ws_admin_uses_stategraph(monkeypatch) -> None:
                 "name": "LangGraph",
                 "data": {"output": {"final_response": "Fake Graph Admin WS Result"}},
             }
+
+        async def astream(self, *args, **kwargs):
+            # ISS-STREAM-002: المسار الحي يستهلك astream(stream_mode=["custom","updates"])
+            yield ("updates", {"synthesizer": {"final_response": "Fake Graph Admin WS Result"}})
 
     def _fake_create_unified_graph():
         return FakeGraph()
@@ -176,7 +196,9 @@ def test_chat_ws_admin_uses_stategraph(monkeypatch) -> None:
     async def fake_persist_assistant_message(**kwargs):
         pass
 
-    monkeypatch.setattr(routes, "_persist_assistant_message", fake_persist_assistant_message)
+    monkeypatch.setattr(
+        chat_stream_engine, "_persist_assistant_message", fake_persist_assistant_message
+    )
 
     token = jwt.encode(
         {"sub": "1", "user_id": 1, "role": "admin", "is_admin": True},
@@ -233,7 +255,9 @@ def test_chat_ws_admin_sanitizes_streaming_errors(monkeypatch) -> None:
         return None
 
     monkeypatch.setattr(routes, "_ensure_conversation", fake_ensure_conversation)
-    monkeypatch.setattr(routes, "_persist_assistant_message", fake_persist_assistant_message)
+    monkeypatch.setattr(
+        chat_stream_engine, "_persist_assistant_message", fake_persist_assistant_message
+    )
 
     token = jwt.encode(
         {"sub": "1", "user_id": 1, "role": "admin", "is_admin": True},
